Adding a shims.json file to keep track of individual spec-compliant ES5 shims.
The root object will contain all globals. Prototype methods will be underneath a "prototype" property.

No shims should auto-shim or modify the environment, but all of them should have an easy mechanism to opt-in to doing so - for example, `object-keys` has a "shim" method for this.

Shims may use UMD if they want, but must use node-style CommonJS modules, and must work with `browserify`.